miku: qcom: Permit overriding TRINKET, STEPPE, NILE from board trees
TRINKET, STEPPE and NILE are Qualcomm board names, not SoC names:
some developers around here prefer overriding these environment
variables to reflect SoC names instead, since that's the usual
way of doing things (for example, in Qualcomm HALs, we have always
used SoC names and NOT board names).

In this case, let the developers decide the names for the platform
that they are supporting and - if they don't choose a preference -
then it's gonna be auto-defined from here.

This is actually useful for Sony devices, in this moment: for
example, we use sm8150 for NILE, sm6125 for TRINKET.

Test: HALs can now be built on Sony SM6125 platforms.

Change-Id: Ic577bf4eeb71d6b3349e5f09991c5755ffc098ed